ERP_Analysis/requirements.txt
pyodbc
pandas
openpyxl
python-dotenv
plotly
kaleido
